Updated cron notification task. Added README and design document.
Updated cron notification task to only send a reminder to users with
active games, as well as reduced the notification interval to a more
sane time. The readme was written to document all of the endpoints and
give details of the structure of the application. A design document
answering prompts given was also added.